pyproject.toml
[tool.poetry]
name = "document-slammer"
version = "0.1.0"
description = ""
authors = ["Raymond F. Wells <rfw2nd@gmail.com>"]
packages = [{include = "document_slammer"}]

[tool.poetry.dependencies]
python = "^3.11"
Jinja2 = "^3.1.2"
Pyright = "^1.1.318"
jsonschema = "^4.18.4"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
pyflakes = "^3.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
